requirements.txt
fastapi~=0.115.5
neo4j~=5.27.0
langchain~=0.3.9
PyYAML~=6.0.2
pydantic~=2.10.2
tqdm~=4.67.1
ollama~=0.4.2
transformers~=4.46.3
typing_extensions~=4.12.2
python-dotenv~=1.0.1